Moved not-found pages to pages folder, removed not needed divs, switched to more suitable alt name for 404 image and increased h1 size
