+++ src/app/app.module.ts
@@ -11,8 +11,8 @@ import { RouterModule } from '@angular/router';
 import { RoutesModule } from './routes/routes.module';
 import { CommonComponentsModule } from './common/components/common-components.module';
 import { HttpClientModule } from '@angular/common/http';
-import { PagesHomePageComponent } from './pages/common/pages-home-page/pages-home-page.component';
-import { PagesNotFoundPageComponent } from './pages/common/pages-not-found-page/pages-not-found-page.component';
+import { PagesHomePageComponent } from './pages/home/pages-home-page.component';
+import { PagesNotFoundPageComponent } from './pages/not-found/pages-not-found-page.component';
 
 @NgModule({
   declarations: [ AppComponent, PagesHomePageComponent, PagesNotFoundPageComponent ],
