--- src/app/pages/common/pages-not-found-page/pages-not-found-page.component.html
@@ -1,9 +1,0 @@
-<div class="container not-found">
-    <div class="app-logo-div">
-        <img class="app-logo" src="assets/images/logos/404-not-found-logo.svg" alt="404-Not-Found-Page-Logo"/>
-    </div>
-    <div class="heading-texts">
-        <h1>404</h1>
-        <p>There is no such resource or page</p>
-    </div>
-</div>
