--- src/app/pages/common/pages-not-found-page/pages-not-found-page.component.scss
@@ -1,34 +1,0 @@
-@use 'sizes';
-
-.not-found {
-    min-height: calc(100vh - #{sizes.$navbar-height});
-    padding-top: 80px;
-    
-    @media only screen and (min-width: 900px) {
-        padding-top: 0px;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-    }
-}
-
-@media only screen and (min-width: 900px) {
-    h1 {
-        font-size: 40px;
-        padding-bottom: 10px;
-    }
-    
-    p {
-        font-size: 26px;
-    }
-}
-
-.app-logo {
-    width: 60%;
-    padding-bottom: 50px;
-}
-
-.app-logo-div, .heading-texts {
-    text-align: center;
-}
