+++ src/app/pages/not-found/pages-not-found-page.component.html
@@ -1,0 +1,5 @@
+<div class="container not-found">
+    <img class="app-logo" src="assets/images/logos/404-not-found-logo.svg" alt="Image for 404 page"/>
+    <h1>404</h1>
+    <p>There is no such resource or page</p>
+</div>
