+++ src/app/pages/not-found/pages-not-found-page.component.scss
@@ -1,0 +1,29 @@
+@use 'sizes';
+
+.not-found {
+    min-height: calc(100vh - #{sizes.$navbar-height});
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    padding-top: 80px;
+
+    
+    @media only screen and (min-width: 900px) {
+        padding-top: 0px;
+        justify-content: center;
+    }
+}
+
+h1 {
+    font-size: 80px;
+    padding-bottom: 25px;
+}
+
+p {
+    font-size: 26px;
+}
+
+.app-logo {
+    width: 30%;
+    padding-bottom: 50px;
+}
