+++ src/app/routes/routes.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { PagesHomePageComponent } from '../pages/common/pages-home-page/pages-home-page.component';
-import { PagesNotFoundPageComponent } from '../pages/common/pages-not-found-page/pages-not-found-page.component';
+import { PagesHomePageComponent } from '../pages/home/pages-home-page.component';
+import { PagesNotFoundPageComponent } from '../pages/not-found/pages-not-found-page.component';
 
 const routes: Routes = [
   { path: '', component: PagesHomePageComponent },
